+++ gantry/util.py
@@ -326,6 +326,7 @@ def build_experiment_spec(
     leader_selection: bool = False,
     host_networking: bool = False,
     propagate_failure: Optional[bool] = None,
+    synchronized_start_timeout: Optional[str] = None,
     mounts: Optional[List[Tuple[str, str]]] = None,
     hostnames: Optional[List[str]] = None,
 ):
@@ -343,6 +344,7 @@ def build_experiment_spec(
             leader_selection=leader_selection,
             host_networking=host_networking,
             propagate_failure=propagate_failure,
+            synchronized_start_timeout=synchronized_start_timeout,
         )
         .with_env_var(name="GANTRY_VERSION", value=VERSION)
         .with_env_var(name="GITHUB_REPO", value=f"{github_account}/{github_repo}")